Make -Wcompat the package default
To be honest, I have no idea what the best practice is
for this stuff. The main advantage is that I don't need
to remember to specify -Wcompat when building locally.
And I think as long as we don't put -Werror there too,
the worst case is more spam while building the package
for the user.